refactor: api 관련 테스트 메서드를 ApiUtils 로 분리
