+++ src/test/java/com/gugucon/shopping/integration/AuthIntegrationTest.java
@@ -2,7 +2,6 @@ package com.gugucon.shopping.integration;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.gugucon.shopping.TestUtils;
 import com.gugucon.shopping.common.exception.ErrorCode;
 import com.gugucon.shopping.common.exception.ErrorResponse;
 import com.gugucon.shopping.integration.config.IntegrationTest;
@@ -10,6 +9,7 @@ import com.gugucon.shopping.item.dto.request.CartItemInsertRequest;
 import com.gugucon.shopping.member.dto.request.LoginRequest;
 import com.gugucon.shopping.member.dto.request.SignupRequest;
 import com.gugucon.shopping.member.repository.MemberRepository;
+import com.gugucon.shopping.utils.ApiUtils;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
@@ -39,7 +39,7 @@ class AuthIntegrationTest {
         final String email = "test_email@woowafriends.com";
         final String password = "test_password!";
         signUp(email, password);
-        String accessToken = TestUtils.login(new LoginRequest(email, password));
+        String accessToken = ApiUtils.login(new LoginRequest(email, password));
 
         // when
         final ExtractableResponse<Response> response = RestAssured
@@ -80,6 +80,6 @@ class AuthIntegrationTest {
 
     private void signUp(final String email, final String password) {
         final SignupRequest request = new SignupRequest(email, password, password,"testUser");
-        TestUtils.signup(request);
+        ApiUtils.signup(request);
     }
 }
